Split ServerConfig and ClientConfig across multiple files
